Initial check-in of pfs-stress tool
The intention of this tool is to exercise the write, overwrite, and
read of files using variable extent sizes. In the process, the read
data will be validated against what was written so as to test for
any data corruption that results.

The current state of pfs-stress is that the .conf file is defined
and the framework for running the multi-threaded test is complete.
What remains is to implement the actual fileStresser() function that
operates on a particular file doing the prescribed operations.