Implemented pfs-stress and updated default .conf file for it
As is evident in pfs-stress.conf, the specified FilePathPrefix
is _pfs_stress_file_... hence, a file path "local" to where
the pfs-stress program is launched.

As a simple modification of the .conf file's "defaults", just
pass the specified path as an "override" on the command line.
For instance, if the path should reference the SMB mount point
of a ProxyFS share at /mnt/smb_proxyfs_mount, either "cd" to
that directory or override FilePathPrefix as in the following:

  pfs-stress pfs-stress/pfs-stress.conf StressParameters.FilePathPrefix=/mnt/smb_proxyfs_mount/_pfs_stress_file_