Checkin a framework for ProxyFS to retry Swift requests that fail. Modify the swiftclient.ObjectGet() method to use the framework as an example.
swiftclient/retry.go defines the RetryCtrl "object" that can be used
to control sleep and retry for a Swift request, including exponential
backoff (the exact values of initial delay, number of attempts, and
exponential backoff paramater still need to be determined).

Add the blunder.TimedOut FSError value which can be returned when all
of the retries fail.

Modify the ObjectGet() method to invoke a new method, objectGetWithRetry()
that wraps the old objectGet() method with a sleep and retry loop when a
request fails.  This is intended to be a model used for the other 19 or so
Swift requests that ProxyFS uses.

the second mechanism is retry.RequestWithRetry(), which handles the entire
process of retrying a request until it succeeds or hits the attempt limit,
and also supports the notion of errors which are not retriable (immediate
abort -- though I haven't written any code to determine which rules should
fall into that category).  it uses a "stupid compiler trick" to allow a
single method to handle retry of all the various types of request that
might need to be retried.  (Stupid compiler tricks are also known as
subtle features of the language used to demonstrate how clever
the programmer using them is.  in this case the feature is a closure.)

the current delay after attempt i is (init_delay * exp_factor^i).  right
now init_delay is 5 sec and exp_factor is 1.2 so the delay increased
by 20% each attempt, i.e. 5 sec, 6 sec, 7.2, 8.6, 10.4, 12.4, 14.9, ...

Note that the backoff code takes into account the amount of time taken
for the request to fail, so if a requset takes 40 sec to fail (unable
to reach a host, say) and the initial backoff is 30 sec then the next
attempt will be made immediately, not 70 sec after the start of the
first requeset).