Add RetryExpBackoff to global config variables and start using it along with RetryDelay and RetryLimit  for the RetryCtrl object.  Adding sanity check for the variables that control Retry.
Add log messages for the cases where a retry succeeds and all retries
fail.  Rework the RequestWithRetry() interface to be simpler and cleaner.

Apparently all config variables are mandatory so I had to populate a
number of config "*_test.go" with the new RetryExpBackoff value or the
tests would fail or even core dump.