Strip out Retry logic that we won't be using.
Add statistics to track number of retry operations (not the number of retries)
and the number of times that retrying succeeded (where initial attempt failed).
RetryRequest() now updates these as appropriate.

Add sanity checking to config for SwiftClient.Retry* paramaters.

Clean up retry.RetryRequest() and better accomodate the case where the
admin has configured SwiftClient.RetryLimit == 0 (no retries).

Insure that RetryRequest() logs a message about "retries failed" even if
retries are disallowed and that it logs a message if the error is not
retriable so that log messages are reasonably consistent with the case
where retries are enabled (SwiftClient.RetryLimit > 0).  In other words,
so that grep'ing on "retry.RequestWithRetry" Error log messages finds
failures even if retries are disabled.

Rework RetryWait() to move the functionality that increments
RetryCtrl.attemptCnt out of RetryWait() and into RetryRequest()