Strip out Retry logic that we won't be using.
Add statistics to track number of retry operations (not the number of retries)
and the number of times that retrying succeeded (where initial attempt failed).
RetryRequest() now updates these as appropriate.

Add sanity checking to config for SwiftClient.Retry* paramaters.

Clean up retry.RetryRequest() and better accomodate the case where the
admin has configured SwiftClient.RetryLimit == 0 (retries disabled).

Insure that RetryRequest() logs a message about "retries failed" even if
retries are disabled; also insure that RetryRequest() logs a message if
the error is not retriable so that log messages are reasonably consistent
with the case where retries are enabled (SwiftClient.RetryLimit > 0).
In other words, make sure that grep'ing on "retry.RequestWithRetry"
Error log messages finds true failures even if retries are disabled.

Rework RetryWait() to move the functionality that increments
RetryCtrl.attemptCnt out of RetryWait() and into RetryRequest()

Rename New() --> NewRetryCtrl()

Develop units tests for RetryCtrl objects and swiftclient.RequestWithRetry().

Add the elapsed time in messages about retries that succeed or fail.

Change values for SwiftClient.RetryLimit and SwiftClient.RetryDelay used
in most tests to be the same as in the python swift client.

Add RetryLimitObject, RetryDelayObject, and RetryExpBackoffObject config
variables.  RetryExpBackoffObject is 2.0 whereas RetryExpBackoff is
changed to 1.2.  The *Object variants apply to object operations while
the others apply to Container and account operations.

Use smaller values for ProxyFs/inode tests so it completes in a reasonable
period of time (6 sec instead of 260 sec).

Initial verson of using the RetryCtrl object for all outstanding requests
Add matching statistics as well.

ObjectDeleteAsync() does not need a retry mechanism.

Add ObjectFetchChunkedPutContext() to the set of
operations with retry handlers.  Have objectCopy() call the
objectFetchChunkedPutContextWithRetry() version.

Lower default timeout from 1 sec to 250 msec for the swiftclient unit tests.

Add code to testRetrySuccess()/Fail() to wait for the stats to update if
they seem to be incorrect.