Implemented package httpserver support for SIGHUP
Note that we also now prevent FSCK from being run on a node other
than the Volume's "owner" (i.e. PrimaryPeer). This prevents corruption
(or, at least, false-positives) from occurring if a node other than
the Volume's "owner" is doing an FSCK while the Volume's "owner" is live.

This check-in will serve as a template for the remaining 9 package
SIGHUP support work now underway.