Modify our logger so it is possible to add additional log targets, where each receives a copy of each log message.  The function logger.AddLogTarget() is used to register a new io.writer function which is called for each log entry.
Define a new object, LogTarget, which is a new logging target that records
the last nEntries log entries in an array that can be examined.

This is useful for testing the AddLogTarget() feature and should also
be a good utility for other unit tests that want to verify output written
to a log.

Fix a couple of errors in TraceEnter()/TraceExit() that affected the
the arguments that were printed (missing whitespace, and "argsprefix"
was not printed if there were no arguments.

Improve the logger unit tests to actually use most of our logger
functionality and to enable trace functions so their output can be seen.