The ProxyFS/inode tests perform an async delete of log segment objects and expect it to happen immediately, which is not a good assumption.
With the new retry logic that causes operations on missing objects to
be retried until we hit the timeout (~31 sec by default) this can take
even longer (i suspect there's a delete of an object that doesn't exist
in the queue).

Check in a temporary(?) fix to decrease the timeouts for retry operations
as well as a permanent fix to wait awhile to see if the ObjectDeleteAsync()
completes.