Removed references to PhysicalContainerLayoutList
Until such time as logic is implemented to support a policy-based
selection among such a list, only the DefaultPhysicalContainerLayout
will be used. As such, the code requiring this currently single-element
list is removed.

In order to enable support for such a list in the future, the code
remains that still verifies that the list does not contain conflicting
elements (e.g. though this is not technically impossible, two layouts
that reference the same ContainerStoragePolicyIndex).