Fix the ObjectChunkedPut retry functionality so it works and write unit tests. (#19)
Updated the comments in api.go for the ChunkedPutContext interface to explain
requirements for its use.

Formerly the ChunkedPutContext interface had a SendChunk() and Close() methods
where SendChunk() would return an error when it encountered a problem and
the Close() method would then retry the entire chunked put that was received
and return an error indicating success or failure.  (There was also a Retry()
method that would do the same thing.)

There were two problems with this.  First, consumers of the ChunkedPutContext
interface immediately abort when SendChunk() return an error and never call
Close() or Retry() (this also means a TCP connection would be leaked).

Second, even if Close() was invoked, a successful retry would only resend
the data passed to ChunkedPutContext upto the point of the error, and not
any other data that was supposed to be part of the object!  So the object
was still effectively corrupt.

Changed the way ChunkedPutContext works so that in "normal" error cases
SendChunk() still returns nil (success) and stashes the data to be sent
away so that the retry can attempt to send it.  Whence Close() is called,
it notices the pending error in ChunkedPutContext.err retries the entire
operation.  If the retry works, it returns success.  Otherwise, it returns
the final error it encountered.

There are some cases where SendChunk() cannot stash away the data to be
sent due to logic errors in the program or corruption of in memory data
structures.  These should probably just be dealt with by panic'ing, but
instead ChunkedPutContext treats this as a "fatal" error, which it returns
to the caller of SendChunk().  If Close() is called later, it does not
attempt to retry but instead returns the same error.  Just in case Close()
is not called (current code does not call Close() after SendChunk()
returns an error, SendChunk() also cleans up the TCP connection.