Implement proper garbage collection in package headhunter
Previously, the means of garbage collecting the on disk portion of
the headhunter's checkpoints involved periodically (e.g. once every
100 checkpoints that, themselves, only occur once every 10 seconds)
taking a "full" snapshot (i.e. by dirtying all the pages of the
B+Trees and flushing them). After this step, all prior objects in
the checkpoint container could be removed. This "stop the world"
approach (since the headhunter database is locked during the checkpoint
operation) is not scalable.

This fix moves to the "on-the-fly" deletion of no longer referenced
objects at each checkpoint. By tracking the number of referenced bytes
in each object, deletions are issued when this count goes to zero.

Further work will eliminate the remaining drivers that force the entire
headhunter database to be in memory.