Fixed swiftclient.ObjectTail() to only return requested []byte buf
Previously, it was also returning the total object size - or so the
comments claimed - but we do not need that piece of information. On
the contrary, the length of the returned []byte buf being different
than what was requested is the only pertinent indication necessary
for when the total object was smaller than this size.