Implemented headhunter.getCheckpoint()
This now upgrades V1 checkpoint info to V2 status
The corresponding putCheckpoint() code is yet to be implemented
Then, finally, the actual taking of a checkpoint logic needs to
be updated to keep the new sortedmap.LayoutReport's up-to-date
as nodes of the B+Trees are allocated and discarded... and to
serialize these as part of each checkpoint. As objects become
unreferenced (bytesUsed == 0), they can be deleted. Hence, no
more "every once in a while" compaction is ever needed anymore.