Feature hhgc - proper garbage collection of headhunter objects (#21)
* Implement proper garbage collection in package headhunter

Previously, the means of garbage collecting the on disk portion of
the headhunter's checkpoints involved periodically (e.g. once every
100 checkpoints that, themselves, only occur once every 10 seconds)
taking a "full" snapshot (i.e. by dirtying all the pages of the
B+Trees and flushing them). After this step, all prior objects in
the checkpoint container could be removed. This "stop the world"
approach (since the headhunter database is locked during the checkpoint
operation) is not scalable.

This fix moves to the "on-the-fly" deletion of no longer referenced
objects at each checkpoint. By tracking the number of referenced bytes
in each object, deletions are issued when this count goes to zero.

Further work will eliminate the remaining drivers that force the entire
headhunter database to be in memory.

* Initial checkpoint loading framework/TODOs to support new V2 format

* Fixed swiftclient.ObjectTail() to only return requested []byte buf

Previously, it was also returning the total object size - or so the
comments claimed - but we do not need that piece of information. On
the contrary, the length of the returned []byte buf being different
than what was requested is the only pertinent indication necessary
for when the total object was smaller than this size.

* Implemented headhunter.getCheckpoint()

This now upgrades V1 checkpoint info to V2 status
The corresponding putCheckpoint() code is yet to be implemented
Then, finally, the actual taking of a checkpoint logic needs to
be updated to keep the new sortedmap.LayoutReport's up-to-date
as nodes of the B+Trees are allocated and discarded... and to
serialize these as part of each checkpoint. As objects become
unreferenced (bytesUsed == 0), they can be deleted. Hence, no
more "every once in a while" compaction is ever needed anymore.

* Reworked how checkpointing & B+Tree management works in headhunter

TODO:
  update addVolume() to utilize getCheckpoint()
  implement putCheckpoint()
  rework checkpointDaemon() to utilize putCheckpoint()
  repair PutNode() to coordinate with putCheckpoint()... and update LayoutReport
  update DiscardNode() to update LayoutReport
  FINALLY - swiftclient.ObjectDeleteAsync() where LayoutReport[] bytesUsed == 0

* Completed get/put of headhunter checkpointing

TODO:
  update addVolume() to utilize getCheckpoint()
  repair PutNode() to update LayoutReport
  update DiscardNode() to update LayoutReport
  FINALLY - swiftclient.ObjectDeleteAsync() where LayoutReport[] bytesUsed == 0

* PutNode() & DiscardNode() now properly garbage collect headhunter DB

TODO: Update addVolume() to utilize getCheckpoint()

* Garbage collection in package headhunter now code complete

* Unit tests past... two system tests fail

  go test github.com/swiftstack/ProxyFS/inode
  go test github.com/swiftstack/ProxyFS/fs

These failures have to do with the sync.WaitGroup getting more Done()
calls than it's had Add() calls...

* Fixed checkpointDoneWaitGroup logic

* Fixed flush of headhunter B+Tree-Object B+Tree

This and another uninitialized couple of variables prevented re-mount
of an existing volume.

* Completed all remaining garbage collection steps in headhunter