Fixed issue where fileInode.size not flushed unless explicitly done
Both Samba (VFS) and NFS neglect to send ProxyFS actual close()
or flush() calls... such that, if we fail or shutdown, the most
recent modifications to a fileInode (write data) will have made
it to Swift... but the metadata for the fileInode will never have
been flushed. This is readily visible when writing a non-zero amount
of data to a new file, rebooting, and observing that the file now
appears to have length zero.

The resolution was to move the time-based flush logic up from
package inode to package fs where appropriate DLM locks can be
taken out to serialize these time-based flushes with other
activity involving the fileInode.