Updated release notes prior to 5.7 release (& hotfix)
Key issue addressed in this release is the lack of fileInode
flushes. The file write data flushes were happening as
expected, but the metadata was not flushed unless explicitly
done so or via another non-R/W metadata modifying operation
was requested.

This release ensures that the implicit (e.g. time-based)
flushes include metadata.