Add "proxyfsd is starting up/shutting down" messages to log.
Add messages that proxyfsd is starting up and shutting down that also
print the proxyfs process PID  (including more messaging about why).

Rearrange the ordering of "daemons" in startup and shutdown in so that
"logger" comes before "stats" in the starting of daemons and after when
they're stopped.

The logger is changed so it does not update statistics about the number
of messages logged if the logger has not started, or has already stopped.

If there's a failure during reconfig (processing a SIGHUP for
configuration update) insure that the error message gets logged.

Unrelated: if inode.fetchOnDiskInode() decides to panic about an inode
fetch, insure the message includes the complete stack.