Updated confMap in httpserver, fixed jrpcfs.Down(), added SIGHUP test
As part of SIGHUP handling, the resulting confMap should be reportable
via the embedded httpserver "config" page. It was not, despite the
walking of the new confMap in response to SIGHUP. The config page now
captures the new confMap at the last moment of SIGHUP handling providing
positive evidence to HTTP clients of the completion of the SIGHUP
handling (a valid assumption since package httpserver is the last
package to process the new confMap).

Previously, as package jrpcfs.Down() executed, it left the goroutines
looping on its two TCP Ports doing Accepts. This leads to Bind errors
when ProxyFS is restarted quickly (such as is done in the recently
added package proxyfsd tests. As part of adding the SIGHUP test to
package proxyfsd, it became evident that ProxyFS was not cleanly
stopping... and left those two TCP Ports bound (for a time).

This set of changes:
1) properly reports new confMap following a SIGHUP
2) closes the Listener sockets as part of jrpcfs.Down()
3) adds SIGHUP testing to package proxyfsd