Added new mkproxyfs cmdline tool
Ultimately, package headhunter will be modified to require
a Volume/Account to be previously formatted. This primarily
avoids inadvertant/silent corruption of pre-existing Swift
Accounts but also dramatically improves Volume mount time.

To that end, mkproxyfs is both a program and a package... the
package there to provide an API for use by test programs in
other packages that need an existing Account to mount.

TODO 1: Currently, in order to not break workflows, the
        headhunter.Up() func still auto-formats... but there
        is a new bool argument to addVolume, allowFormat,
        that will ultimately be changed to doFormat... and
        will not auto-format if not specified.

TODO 2: As part of the workflow change (i.e. where headhunter.Up()
        will no longer auto-format), various package tests and
        regression_test.py will need to be modified to call the
        `mkproxyfs -N` tool (or, see TODO 3, mkproxyfs.Format()).

TODO 3: Package mkproxyfs could use a unit test... indeed, that
        the initial motivation for making mkproxyfs a package
        anyway (i.e. so that the test could invoke
        mkproxyfs.Format() for each test case).