Correct inode timestamp updates and permission checks
Moidyf link() so we don't risk deadlock if a client tries to make a
hardlink to a directory (getting locks out of order).  Don't put
error messages in the log when the client sends us an invalid request
that could be legitmately generated (like a request to make a hardlink to
a directory).

Various file system operations were failing to update ctime (attribtute
change time) when they should and updated atime (access time) when
they shouldn't.

Correct one permission checking case -- even the owner of a file needs
write permission to change its size (truncate(2)).  Flagged with
"TODO ... --craig" a number of other cases where our checking seems
incorrect.

There were a few places where we updated inode timestamps but may not
have marked the inode dirty. I added:
    inode.dirty = true
which should be benign if its also being set somewhere else.

Flagged with "TODO ... --craig" a place where we seem to remove an
inode from the flush list when we shouldn't.

rewrite checkMetadata() so that it catches field names that don't
exist (typos in the test cases).