Add RPC to create a directory by path
We've got jrpcfs.RpcMkdir, but pfs_middleware can't use it because it
needs mount IDs and inode numbers and such. Now there's also
jrpcfs.RpcMiddlewareMkdir that takes a path and some metadata, and
those are things that pfs_middleware has handy.

The goal is to let an object PUT request with Content-Length=0 and
Content-Type=application/directory create a directory, not a zero-byte
file. (The *real* goal is functioning S3 multipart uploads with
swift3, but this is one of many necessary steps.)