fix hard link err message logged by addDirEntryInMemory()
volumeStruct.Link() is called to implement both file create (Create())
and hard link (Link()), and it calls addDirEntryInMemory() to do the real
work.  addDirEntryInMemory() logs an error if the filename already exists.
change it to stop logging any message for something that can be caused
by client request, and amend the error message returned to the caller to
say the "link" failed beccause the entry exists.

in mountStruct.Create() push the creation of the new inode down to
just before we need so it doesn't need to be destroyed if one of the
three checks performed earlier fails.