First half of fix for panic in flush code.
in fs.Unlink(), remove the correct inode from the inFlightFileInodeDataTracker()
map (remove the file inode, not the parent directory inode).

in inFlightFileInodeDataTracker() don't panic() if there's a problem flushing
the inode to disk (in Flush()), just log an error and go on with life.

Convert other calls to panic() in that routine to calls to logger.PanicfWithError()
so we get more info in the log.

Add trace messages in various places to track down this problem.

Cleanup the way we print inode numbers (decimal instead of a 16 character
hex string).

Add code so its possible to enable tracing in some packages that did not
support it (dlm, fs, fuse, headhunter, proxyfsd, swiftclient), not that
I added trace statements to all of those.