Clear out hung FUSE Mount Point on restart
Method chosen it so issue (lazy) unmount and await its completion
if it is found that the FUSE Mount Point is still active. This is
assessed by recognizing when the Device for the FUSE Mount Point
and its containing directory differ. This same property is now used
to verify when the FUSE Mount Point is available.