Fix the "called while active" bug in SendChunk() retries.
This is part of the Chunked Put stuff.

There's a bug in the handling of retries for SendChunk() which can be hit
after an initial request fails and then the first retry fails as well.
It logs the error:

     swiftclient.chunkedPutContext.retry() called while active

The problem is that we were not clearing the active flag in
chunkedPutContext after a retry failed, so the next retry will see the
flag set and abort.  It turns out that in this case we also fail to tear
down the TCP connection and put it back in the pool of connections,
meaning proxyfsd could slowly leak file descriptors.

Also added a counter, sendChunkRetryCnt, to swiftclient/object.go that
allows us to simulate errors in the retry function.  Verify that the
unit tests hit the same bug and then fixed it.

(I only found the second bug after investigating why the swiftclient test
would occasionally take 133 sec to run instead of 13.)

Clean up the logged error messages to make them shorter and more accurate.