Make ETags MD5 hashes sometimes
After initial PUT via Swift API, an object's ETag will be its MD5
hash. If the object is written to, then its ETag will be the synthetic
one ("pfsv2/..."). This applies to both object GET/HEAD and to
container GET.

Also add a "-32" to our synthetic ETag and change version from "pfsv1"
to "pfsv2". This may help with some S3 clients who currently see a
synthetic ETag and become unable to download any object. In S3, ETags
for multipart objects end in -N, where N is the number of segments; by
adding -32 to ours, we might get some S3 clients to skip MD5
comparison.

Made the S3Compat middleware harmless. It's not done yet, but it was
crashing due to a call to a nonexistent function.