Make ETags MD5 hashes sometimes
After initial PUT via Swift API, an object's ETag will be its MD5
hash. If the object is written to, then its ETag will be the synthetic
one ("pfsv2/...").

Also add a "-32" to our synthetic ETag and change version from "pfsv1"
to "pfsv2". This may help with some S3 clients who currently see a
synthetic ETag and become unable to download any object.

Made the S3Compat middleware harmless. It's not done yet, but it was
crashing due to a call to a nonexistent function.