Add "statslogger" to ProxyFS.
statslogger periodically logs selected statistics on free Swift
connections, memory usage, and Swift operations.  memory usage
and Swift operations are logged as both an absolute value (total
since startup) and as a delta since the last statistics were
logged.

the log interval is controlled by a new config variable
StatsLogger.Period, which can be set as low as 1 sec (use 0 to disable
logging of statistics).  It defaults to 10 min, but logs a warning
message when it uses the default.