Up'd default SwiftClient to be more realistic (matching SSController defaults)
The previous value of 64 for the chunked PUT pool was woefully unrealistic...
It basically means that a ProxyFS node can only support 64 simultaneously
written files... plenty for development... completely impractical for the
real world.

This is only a temporary fix... the real fix would endeavor to trigger
flushes on other file write streams to "make room" for a new one to
start (with some limits, of course - we don't want a bunch of one byte
objects out there).