Add ability for ProxyFS to set Account-level SysMeta header
This required adding the "meta" middleware filter to the NoAuth pipeline.
The "meta" middleware allows the NoAuth client (i.e. ProxyFS) to set the
"X-ProxyFS-BiModal" Header... As this will be used on a Swift Account,
the "meta" middleware translate this header name (value unaffected) to
"X-Account-Sysmeta-Proxyfs-Bimodal". Meanwhile, the "pfs" middleware,
sitting after the "gatekeeper" middleware, will be able to see this header
on the BiModal Accounts that ProxyFS has indicated... and, of course, not
on non-B-Modal Accounts.

TODO:
  All systems need to deploy the updated NoAuth pipeline to include "meta"
  ProxyFS, when formatting a Swift Account, needs to:
    POST...-H "X-ProxyFS-BiModal: true"
  PFS middleware needs to HEAD every Swift Account looking for:
    "X-Account-Sysmeta-Proxyfs-Bimodal" header (value unimportant)