Fix a NULL pointer reference in proxyfs.
During a flush by headhunter, if the chunked put context (the put to an
object) exceeds a maximum size then it is closed and new one opened.
closeCheckpointChunkedPutContextIfNecessary() checks for the size and
closes it.  However, if it does close the chunked put context, it doesn't
nil out the pointer 'volume.checkpointChunkedPutContext', so headhunter
thinks the chunked put context is still open and tries to use it,
causing this problem.

So set the pointer to nil when the chunked put context is closed.

closeCheckpointChunkedPutContext() has a similar but where it doesn't
nil the pointer if Close() returns an error.  fix that as well.

Fix chunkedPutContextStruct.Close() so it won't trigger a SIGSEGV in
the situation where its handed an inactive (closed) chunked put context.

Really I need to change Close() to panic in this situation, but I'm
going to do that using logger.PanicWithError() which needs some work
before its ready for that.