Add the trackedlock package to watch over locks.
trackedlock.Mutex can wrap sync.Mutex to provide the same interface
but also watches to see how long locks are held and writes warnings
to the log if they are locked too long.

trackedlock.RWMutex does the same for sync.RWMutex.

Expose the interfaces to the lock tracking code so that DLM locks, and
potentially the chunked put connection code can use them for tracking.

Partial fix for trackedlock if locked RWMutex transferred:

    If the goroutine that calls RUnlock() is different then the one that
    calls RLock(), the lock tracking can draw incorrect conclusions (for
    example, it can't figure out which of several RLock() operations is
    being unlocked so it can't know if one held the lock for too long) .

    Add code to cut down on the pain of this issue (but its not a
    complete fix).

    A proper fix would require that a goroutine that is passing a locked
    RWMutex (or DLM lock) to another goroutine (or the gorouting the
    receives the lock) notify the trackedlock code of the transferal
    through an API that needs to be defined.  We'll get to that later
    if and when we track lock ownership.

Write a whole bunch of unit tests.

Lock tracking is disabled by default.

Changed trackedlock so storage for stack traces is allocated only when
a stack trace needs to be preserved and use sync.Pool to hold the
objects.

This results in much lower memory consumption for proxyfsd under load.

Change the lockWatcher in the trackedlock to report the 16
locks that were held the longest and not just 1 (controlled by
globals.lockWatcherLocksLogged which is not tunable, but could
be changed via gdb).

Seeing more locks should be useful in debugging deadlocks (its likely
that the lock held second longest is the lock that the thread holding
the first lock wants) and may also help in those cases where the lock
held longest is not the cause of the deadlock (perhaps because a thread
got the lock and did some time-consuming work before it then tried to
acquire the lock held by the thread causing the deadlock.

Also try to minimize the chance of hitting a SIGSEGV that can occur if
we try to copy the lock stack of a Mutex that has been held too long
and it happens to be unlocked right then.

The change to log multiple locks held too long required changes in the
tests to look at the new "rank %d" log field.

Tighten up the log entry parser logger.ParseLogEntry() a bit.  Have it
log a warning if a log entyr doesn't match its RE.  Fix some bugs in
swiftclient tests that were exposed by this.

Improve comments in this and related code.  Improve the trackedlock
log messages as well.