Updated package fs SIGHUP/unmount logic to flush *ALL* volumes
Previously, only volumes being unmounted would be flushed. Alas, it is
possible that SIGHUP is being sent to proxyfsd to indicate is should
(also) reconnect with the NoAuth Swift Proxy. This necessitates all
connections must be restarted... and the pre-SIGHUP, whether or not
being used by a departing volume, need to be completed/closed.

This change pre-dates a companion change down in package swiftclient
that will handle SIGHUP by re-provisioning the two connection pools
it is maintaining. It can only actually complete this process (in a
timely manner) if layers above swiftclient choose to finish use of
those connections. This is not really an issue for "normal" HTTP
requests, but for the Chunked PUT HTTP requests, these may be very
long lived. Performing a flush in package fs on *ALL* volumes avoids
this problem.