Add ParseLogEntry() and ParseLogForFunc() to logger.
Move some code from parseRetryLogEntry() in the swiftclient unit tests
to logger and rename it PraseLogEntry().  Add code to handle embedded
double quote ('"') in msgs or errors and also add unit tests.

Add ParseLogForFunc() parser to logger, which looks through the last
N lines of the log for messages from a particular function name with
a message that matches a provided passed regular expression and return
the fields matched.