Fix the default value for BPlusTreeObjectCacheEvictHighLimit
If BPlusTreeObjectCacheEvictHighLimit is not specified in the
proxyfs config file, the default kicks in:

        bPlusTreeObjectCacheEvictHighLimit, err = confMap.FetchOptionValueUint64("FSGlobals", "BPlusTreeObjectCacheEvict\
HighLimit")
        if nil != err {
                // TODO: eventually, just return
                bPlusTreeObjectCacheEvictHighLimit = 10010
        }

unfortunately, it doesn't reset err to nil.  the next code that
overwrites err is:

        for _, volumeName = range volumeNames {

but if no volumes (file systems) are defined in config then err is not
overwritten with nil and headhunter.Up() returns an error so that
proxyfs cannot start.

reset err for all of the default cases.