Get rid of inode.InodeRootGroupID.
There is no such as that as the "Root Group" in UNIX.  Group ID 0 does
not have any special priviledges.

Get rid of inode.InodeRootGroupID throughout the code and replace it
with inode.InodeGroupID(0).

That's the only change in this commit, but it hits a lot of files which
is why i broke it out as a separate commit.   You don't have to read it
if you're not interested in that.