Lots of changes for permission checking.
Add a new parameter to (*mounstruct).Access(..., accessMode inode.InodeMode) which
is set to OwnerOverride if the owner can override the permission checks (for
example, granting the owner write permission to a file she owns but the 'w' bit
for owner is not set; this is necessary to allow writing to read-only files via
NFS).

Fix other problem with permission checks that i noticed while propagating this
change, e.g. eXecute permission is not required to perform a readdir().

Did not fix some others: write permission is required to get a write lock on
a file, but the code only checks for read permission.

Fix some bugs that Gerard found where we did not perform enough checks when
changing the ownership of a file (only SuperUser, i.e. root can do that), but
the owner is allowed to "change" it to the owner and similar issues with
changing group ownership.