pfs_middleware: show correct Content-Type in container listing
If you make an object PUT request with a particular Content-Type,
Swift will show that Content-Type in object GET/HEAD responses and
also in container GET responses that reference the object.

pfs_middleware had the object GET/HEAD responses right, but was only
guessing the Content-Type from the filename in the container GET
response. This commit makes the behavior match Swift's.