pfs_middleware: fix 500 on some object GETs
If you ask for a path that treats a file like a directory, proxyfsd
returns an error with error-code 20 (NotDirError). We used to not
handle that and return a 500; now we correctly return a 404 in that
case.

Example: given a file stuff/vacation/grand-canyon.jpg, a GET request
for /v1/AUTH_me/stuff/vacation/grand-canyon.jpg/something-else would
trigger this bug.