pfs_middleware: preserve ETag + sysmeta on POST
Object POST handling was not correctly preserving existing metadata
the way Swift does: user metadata is blown away, sysmeta is preserved,
and a few special fields are also preserved. Now we match better.

This should help with the Swift functional tests somewhat, as now POST
requests to SLO or DLO manifests won't destroy the is-a-manifest
property.