Fix XML account GET responses.
In Swift 2.16.0, a new listing_formats middleware was introduced near
the left end of the pipeline. For everything to the right of
listing_formats, all account and container listings are now JSON-only,
and the listing_formats middleware takes care of conversion to XML or
plaintext.

However, since Swift's staticweb middleware can be configured to reply
with JSON, listing_formats has a fallback behavior: if the response it
gets doesn't look like a normal account or container GET response,
then the response is returned as-is.

pfs_middleware's JSON account GET response didn't have the
"last_modified" key for each entry because the RPC wasn't returning
that information. Thus, listing_formats failed to convert it, and so
sent it to the client untranslated.

The net effect was sort of funny: the client said GET
/v1/a?format=xml, and got back JSON.

This commit widens the return value of Server.RpcGetAccount to include
the modification time of each entry, then uses that to include
"last_modified" in each entry in the JSON account GET response.

For compatibility with older Swift versions (i.e. those lacking
listing_formats), this commit also adds a <last_modified> tag to each
entry in pfs_middleware's XML account GET response.